Moved common code from Core.Util.Fst.FST and TestFramework.Util.Fst.FSTTester into the support namespace.
